move additional guests to own component
