+++ components/Form/AdditionalGuests/index.js
@@ -1,0 +1,3 @@
+import AdditionalGuests from './AdditionalGuests';
+
+export default AdditionalGuests;
